requirements.txt
numpy~=1.24.3
pandas~=2.0.3
opencv-python~=4.9.0.80
matplotlib~=3.7.2
tqdm~=4.65.0
pillow~=9.4.0
ipywidgets~=8.0.4
ipython~=8.15.0
simpleitk~=2.3.1
scikit-image~=0.20.0
scipy~=1.11.1
torchvision~=0.17.0
setuptools~=68.0.0
monai~=1.3.0
transformers~=4.32.1
pathlib~=1.0.1